Splits light sdk queries into multiple files
